refactoring #20 - delete baseViewController
